Indent and strip package metadata strings.
The Python package metadata was not being correctly formatted due to the
presence of unindented lines. I think setuptools should be responsible
for this formatting, but it seems easier to just fix it here.

With this change, it becomes possible for the standard pkg_resources
parser to read the PKG-INFO metadata for this project.